--- public/app.js
@@ -1,3 +1,0 @@
-import App from "./component/App.js";
-
-new App(document.querySelector('#root'));
