+++ public/index.html
@@ -11,8 +11,7 @@
     <link rel="stylesheet" href="./css/comment.css" />
     <script src="https://unpkg.com/boxicons@2.1.1/dist/boxicons.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js"></script>
-    <script defer type="module" src="app.js"></script>
-    <!-- <script defer type="module" src="router.js"></script> -->
+    <script defer type="module" src="index.js"></script>
   </head>
   <body>
     <div id="root"></div>
