+++ public/index.js
@@ -1,0 +1,3 @@
+import App from './component/App.js';
+
+new App(document.querySelector('#root'));
